requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
matplotlib
h5py
 
